add timeout for migration deferred response

+++ redis/redis-console/src/main/java/com/ctrip/xpipe/redis/console/config/ConsoleConfig.java
@@ -128,6 +128,6 @@ public interface ConsoleConfig extends CoreConfig, CheckerConfig, AlertConfig {
 
 	int getCheckerAckTimeoutMilli();
 
-	int getMigrationTimeoutMilli();
+	long getMigrationTimeoutMilli();
 
 }
